requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
defusedxml==0.7.1
fonttools==4.60.1
fpdf2==2.8.4
idna==3.10
pillow==11.3.0
python-nmap==0.7.1
requests==2.32.5
urllib3==2.5.0
